feat: Support Chef & Salt
This adds support for Salt's grains and Chef's ohai. Both of these tools
contain "null" values which json.Unmarshal turns into "nil". These nil
values need to be stripped prior to passing to the plist library or
gosal will panic.

I've renamed facter.go to cfgmgmnt.go to properly represent the new
support.

Fix #24
